test(Slippy Map): Tests are passing options
