+++ ui/app/containers/SlippyMap/index.jsx
@@ -15,7 +15,7 @@ const Map = ReactMapboxGl({
   accessToken: 'pk.eyJ1IjoiZGFuYWN0aXZlIiwiYSI6ImNpdmo0OGo2YTAxcGIyenBkZWZlN3Ewam4ifQ.npY0cY_HdHg1OB692HtcUw',
 });
 
-const SlippyMap = ({ currentMemory: { coordinates, coordinateAccuracy }, items }) => (
+const SlippyMap = ({ currentMemory: { coordinates, coordinateAccuracy } = {}, items }) => (
   <Map {...transformMapOptions({ coordinates, coordinateAccuracy })}>
     <Source {...transformSourceOptions({ items })} />
     <Layer {...clusterOptions} />
